Move auth backed to config options
Intial design fixed the 'opensuse' auth backend in base settings, and the
login button was wired directly to it. As part of making the application more
accessible to a larger community, the auth backends needed to move out to
an external configuration option; this also allowed for enabling multiple
backends.