Downloadable credentials
Provide an interface for downloading credentials, which are presented only at
checkout (and not persisted for user security).

Offering the download was complicated by not having persisted credentials, as
downloads typically require an additional request to the server - which no
longer has the credentials context.

The data: URI (https://developer.mozilla.org/en-US/docs/Web/HTTP/data_URIs ) in
addition to the download attribute of HTML links, allows offering a client-side
download from the existing checkout page, without an additonal round-trip to
the server (and the inherent requirement to persist user credentials).

Resolves issue https://github.com/openbare/openbare/issues/12